Setup all three sections to update resume html
The three main sections are implemented and working, meaning their
inputs are being updated live to the mock resume on the other side of
the web page.

Now, for the Education and Experience sections, I need to tweak it so
they are containers for multiple places. For example, for the app user
to be able to list multiple different schools or previous places of
employment.