Update collapsable section functionality
Education and experience sections (collapsable edit sections) now list
out the contents of the objects as buttons. When those buttons are
clicked, it clears the section and opens the edit menu. This menu
includes buttons to save and delete (working on delete functionality),
which close the edit and show the list again.